pyproject.toml
[project]
name= "llm-engineering"
version = "0.1.0"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
